+++ Advent-of-Code-2024/AdventOfCode2024/Day01.cs
@@ -5,7 +5,7 @@ namespace AdventOfCode2024;
 internal class Day01 : BaseDay
 {
 
-    public override int Puzzle1()
+    public override long Puzzle1()
     {
         int answer = 0;
 
@@ -20,7 +20,7 @@ internal class Day01 : BaseDay
         return answer;
     }
 
-    public override int Puzzle2()
+    public override long Puzzle2()
     {
         int answer = 0;
