+++ Advent-of-Code-2024/AdventOfCode2024/Day02.cs
@@ -8,7 +8,7 @@ namespace AdventOfCode2024;
 
 internal class Day02 : BaseDay
 {
-    public override int Puzzle1()
+    public override long Puzzle1()
     {
         int answer = 0;
         var inputs = Input
@@ -28,7 +28,7 @@ internal class Day02 : BaseDay
         return answer;
     }
 
-    public override int Puzzle2()
+    public override long Puzzle2()
     {
         int answer = 0;
         var inputs = Input
