+++ Advent-of-Code-2024/AdventOfCode2024/Day04.cs
@@ -9,7 +9,7 @@ namespace AdventOfCode2024;
 
 internal class Day04 : BaseDay
 {
-    public override int Puzzle1()
+    public override long Puzzle1()
     {
         int answer = 0;
         var input = Input.Split("\r\n");
@@ -22,7 +22,7 @@ internal class Day04 : BaseDay
         return answer;
     }
 
-    public override int Puzzle2()
+    public override long Puzzle2()
     {
         int answer = 0;
         var input = Input.Split("\r\n");
